refactor(backend): split intrinsic helpers and optimize typed iteration - Organize compiler-owned C intrinsic support into focused headers behind an umbrella include - Preserve Array reference semantics while using indexed element access for iteration - Share dictionary key snapshots and cache contiguous key storage across iterator copies - Dispatch packed-array iteration to typed helpers with cached element pointers instead of Variant indexing - Update frontend contracts and runtime documentation for revised iterator ownership and materialization rules
